small nits
