+++ charts/kyverno/templates/crds.yaml
@@ -5,6 +5,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '1'
+    internal.config.kubernetes.io/index: '1'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
@@ -2871,6 +2872,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '2'
+    internal.config.kubernetes.io/index: '2'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
@@ -3142,6 +3144,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '3'
+    internal.config.kubernetes.io/index: '3'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
@@ -3413,6 +3416,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '4'
+    internal.config.kubernetes.io/index: '4'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
@@ -3594,6 +3598,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '5'
+    internal.config.kubernetes.io/index: '5'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
@@ -6460,6 +6465,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '6'
+    internal.config.kubernetes.io/index: '6'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
@@ -6731,6 +6737,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.8.0
     config.kubernetes.io/index: '7'
+    internal.config.kubernetes.io/index: '7'
   creationTimestamp: null
   labels:
     app.kubernetes.io/component: kyverno
